I separate title and logo in header it into separate components, in other I almost immediately remade it on React. Some fixing the css of the list of cocktails and other things. The initial steps to creating a navigation for a cocktail page.
