+++ src/components/cocktails_list/quick_filters_menu/quick_filters_menu.jsx
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import filter_elements_list from '@components/main_page/cocktails_list/quick_filters_menu/quick_filters_menu_elements_list';
+import filter_elements_list from '@components/cocktails_list/quick_filters_menu/quick_filters_menu_elements_list';
 
 const QuickFiltersMenu = () => {
     const [isChecked, setChecksList] = useState(
