+++ src/pages/main_page.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Header from "@components/main_page/header/header";
-import QuickFiltersMenu from "@components/main_page/cocktails_list/quick_filters_menu/quick_filters_menu";
-import CocktailsList from "@components/main_page/cocktails_list/cocktails_list";
-import Footer from "@components/main_page/footer/footer";
+import Header from "@components/header/header";
+import QuickFiltersMenu from "@components/cocktails_list/quick_filters_menu/quick_filters_menu";
+import CocktailsList from "@components/cocktails_list/cocktails_list";
+import Footer from "@components/footer/footer";
 
 const MainBlock = () => {
     return (
